check if font installed

+++ install.sh
@@ -45,22 +45,26 @@ declare -a fonts=(
 version='2.2.2'
 fonts_dir="${HOME}/.local/share/fonts"
 
-if [[ ! -d "$fonts_dir" ]]; then
-	mkdir -p "$fonts_dir"
-fi
+if fc-match "Hack Nerd Font Mono" &>/dev/null; then
+    echo "Hack Nerd Font Mono is already installed."
+else
+	if [[ ! -d "$fonts_dir" ]]; then
+		mkdir -p "$fonts_dir"
+	fi
 
-for font in "${fonts[@]}"; do
-	zip_file="${font}.zip"
-	download_url="https://github.com/ryanoasis/nerd-fonts/releases/download/v${version}/${zip_file}"
-	echo "Downloading $download_url"
-	wget "$download_url"
-	unzip "$zip_file" -d "$fonts_dir" -x "*.txt/*" -x "*.md/*"
-	rm "$zip_file"
-done
+	for font in "${fonts[@]}"; do
+		zip_file="${font}.zip"
+		download_url="https://github.com/ryanoasis/nerd-fonts/releases/download/v${version}/${zip_file}"
+		echo "Downloading $download_url"
+		wget "$download_url"
+		unzip "$zip_file" -d "$fonts_dir" -x "*.txt/*" -x "*.md/*"
+		rm "$zip_file"
+	done
 
-find "$fonts_dir" -name '*Windows Compatible*' -delete
+	find "$fonts_dir" -name '*Windows Compatible*' -delete
 
-fc-cache -fv
+	fc-cache -fv
+if
 
 # Install oh-my-zsh and set zsh as default
 sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" --keep-zshrc
